requirements.txt
eccodes==1.4.1
pandas==1.3.3
progressbar==2.5
numpy==1.19.5
matplotlib==3.4.3
google-api-core==2.8.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
gcsfs==2022.3.0
pdoc3==0.10.0